Task Annotation and TaskRecord
Cant scan for subclasses of Task from a separate jar built with the exported library. Need to use annotations. Possibility to customise the parameters of the @Executer. I should have a Annotation processor for editor warnings about incorrect setup for that annotation. Maybe a better solution is to implement runnable